generator/requirements.txt
Flask==2.0.1
requests==2.25.1
python-dotenv==0.19.0
Flask-SocketIO==5.1.1
Werkzeug==2.0.1
flask-cors==3.0.10
aiohttp==3.7.4.post0
